requirements.txt
amqp==2.6.0
amqplib==1.0.2
anyjson==0.3.3
billiard==3.6.3.0
boto3==1.13.25
botocore==1.16.25
celery==4.4.5
certifi==2020.4.5.1
chardet==3.0.4
dj-stripe==2.3.0
Django==2.2.5
django-celery==3.3.1
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-timezone-field==4.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.10
mysql==0.0.2
mysqlclient==1.4.6
pkg-resources==0.0.0
psycopg2-binary==2.8.4
py3dns==3.2.1
pycurl==7.43.0.5
Pygments==2.4.2
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.2
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
stripe==2.48.0
urllib3==1.25.9
validate-email==1.3
vine==1.3.0
zipp==3.1.0
